+++ DAL/DALPhongBan.cs
@@ -12,14 +12,26 @@ namespace DAL
     public class DALPhongBan
     {
         private readonly string connectionString;
-        private readonly PersonnelManagementDataContextDataContext _dbContextPB;
+        private readonly PersonnelManagementDataContextDataContext _dbContext;
 
         public DALPhongBan(string conn)
         {
             connectionString = conn;
-            _dbContextPB = new PersonnelManagementDataContextDataContext(conn);
+            _dbContext = new PersonnelManagementDataContextDataContext(conn);
         }
 
+        public DTOPhongBan FindPhongBanByIdChucVu(int id)
+        {
+            DTOPhongBan phongBan = (from pb in _dbContext.PhongBans
+                                    join cv in _dbContext.ChucVus on pb.id equals cv.idPhongBan
+                                    select new DTOPhongBan
+                                    {
+                                        Id = pb.id,
+                                        TenPhongBan = pb.TenPhongBan,
+                                        MoTa = pb.Mota
+                                    }).FirstOrDefault();
+            return phongBan;
+        }
         public DataTable GetAllPhongBan()
         {
             using (SqlConnection conn = new SqlConnection(connectionString))
@@ -45,8 +57,14 @@ namespace DAL
             }
         }
 
-        public bool InsertPhongBan(DTOPhongBan pb)
+        public string InsertPhongBan(DTOPhongBan pb)
         {
+            bool checkTenPhongBan = _dbContext.PhongBans.Any(x => x.TenPhongBan == pb.TenPhongBan);
+            if(checkTenPhongBan)
+            {
+                //Tên phòng ban mới đã tồn tại trong database
+                return "Tên phòng ban đã tồn tại không thể thêm";
+            }
             using (SqlConnection conn = new SqlConnection(connectionString))
             {
                 string query = "INSERT INTO PhongBan (TenPhongBan, MoTa) VALUES (@Ten, @MoTa)";
@@ -54,12 +72,22 @@ namespace DAL
                 cmd.Parameters.AddWithValue("@Ten", pb.TenPhongBan);
                 cmd.Parameters.AddWithValue("@MoTa", pb.MoTa ?? (object)DBNull.Value);
                 conn.Open();
-                return cmd.ExecuteNonQuery() > 0;
+                if(cmd.ExecuteNonQuery() > 0)
+                {
+                    return "Thêm phòng ban thành công!";
+                }
+                return "Thêm phòng ban thất bại!";
             }
         }
 
-        public bool UpdatePhongBan(DTOPhongBan pb)
+        public string UpdatePhongBan(DTOPhongBan pb)
         {
+            bool checkTenPhongBan = _dbContext.PhongBans.Any(x => x.TenPhongBan == pb.TenPhongBan && x.id != pb.Id);
+            if (checkTenPhongBan)
+            {
+                //Tên phòng ban mới đã tồn tại trong database
+                return "Tên phòng ban mới đã bị trùng, vui lòng nhập tên khác!";
+            }
             using (SqlConnection conn = new SqlConnection(connectionString))
             {
                 string query = "UPDATE PhongBan SET TenPhongBan=@Ten, MoTa=@MoTa WHERE id=@id";
@@ -68,26 +96,50 @@ namespace DAL
                 cmd.Parameters.AddWithValue("@MoTa", pb.MoTa ?? (object)DBNull.Value);
                 cmd.Parameters.AddWithValue("@id", pb.Id);
                 conn.Open();
-                return cmd.ExecuteNonQuery() > 0;
+                if (cmd.ExecuteNonQuery() > 0)
+                {
+                    return "Cập nhật thông tin phòng ban thành công!";
+                }
+                return "Cập nhật phòng ban thất bại!";
             }
         }
 
-        public bool DeletePhongBan(int id)
+        public string DeletePhongBan(int id)
         {
             using (SqlConnection conn = new SqlConnection(connectionString))
             {
+                var ktraIDPB = _dbContext.ChucVus.FirstOrDefault(cv => cv.idPhongBan == id);
+                var ktraIDNV = _dbContext.NhanViens.FirstOrDefault(cv => cv.idPhongBan == id);
+
+                if (ktraIDPB != null && ktraIDPB != null)
+                {
+                    return "Phòng ban hiện đang có dữ liệu nhân viên và chức vụ, nên không thể xóa";
+                }
+                if(ktraIDPB != null)
+                {
+                    return "Phòng ban hiện đang có dữ liệu chức vụ, nên không thể xóa";
+                }
+                if (ktraIDPB != null)
+                {
+                    return "Phòng ban hiện đang có dữ liệu nhân viên, nên không thể xóa";
+                }
+
                 string query = "DELETE FROM PhongBan WHERE id=@id";
                 SqlCommand cmd = new SqlCommand(query, conn);
                 cmd.Parameters.AddWithValue("@id", id);
                 conn.Open();
-                return cmd.ExecuteNonQuery() > 0;
+                if(cmd.ExecuteNonQuery() > 0)
+                {
+                    return "Xóa phòng ban thành công!";
+                }
+                return "Xóa phòng ban thất bại!";
             }
         }
 
         // Lay ten phong ban
-        public string LayTenPhongBan(int id) => _dbContextPB.PhongBans.Where(p => p.id == id).Select(p => p.TenPhongBan).FirstOrDefault().ToString() ?? string.Empty;
+        public string LayTenPhongBan(int id) => _dbContext.PhongBans.FirstOrDefault(p => p.id == id).TenPhongBan ?? string.Empty;
 
         // Lay ds phong ban
-        public List<PhongBan> LayDsPhongBan() => _dbContextPB.PhongBans.ToList();
+        public List<PhongBan> LayDsPhongBan() => _dbContext.PhongBans.ToList();
     }
 }
